edhoc: use 16 bits to avoid warning about always true comparison

+++ os/net/security/edhoc/edhoc-msgs.c
@@ -456,7 +456,7 @@ edhoc_get_key_id_cred_x(uint8_t **p, uint8_t *out_id_cred_x, cose_key_t *key)
     return -1;
   }
 
-  uint8_t id_cred_x_sz = *p - start;
+  uint16_t id_cred_x_sz = *p - start;
   if(out_id_cred_x != NULL) {
     assert(*p - start >= 0);
     assert(id_cred_x_sz <= MAX_BUFFER);
